refactor how errors are handled by creating a FormHandler class
