+++ app/guest/add_guest.mustache
@@ -4,20 +4,17 @@
   <label for="first_name">
     First Name
     <input type="text" name="first_name" id="first_name" value="" />
-    {{#form.errors.first_name}}
-    <span class="error">{{.}}</span>
-    {{/form.errors.first_name}}
   </label>
 
   <label for="last_name">
     Last Name
     <input type="text" name="last_name" id="last_name" value="" />
-    {{#form.errors.last_name}}
-    <span class="error">{{.}}</span>
-    {{/form.errors.last_name}}
   </label>
 
   {{{ form.csrf_token }}}
+  {{#errors}}
+  <span class="error">{{error}}</span>
+  {{/errors}}
 
   <input type="submit" value="Continue →">
 </form>
