+++ app/guest/views.py
@@ -1,7 +1,8 @@
 from app.views.template import Template
+from app.forms import FormHandler
 
 
-class AddGuest(Template):
+class AddGuest(Template, FormHandler):
 
     def __init__(self, invite_token, form):
         self.token = invite_token
